requirements.txt
discord.py
python-dotenv
motor
g4f
nest_asyncio
yt-dlp
pynacl
pyfiglet
termcolor
google-generativeai